pyproject.toml
[tool.poetry]
name = "collaberr-backend"
version = "0.1.0"
description = ""
authors = ["Tony Lee <43350428+TonysHub@users.noreply.github.com>"]
readme = "README.md"
license = "MIT"


[tool.poetry.dependencies]
argon2-cffi = "^21.3.0"
black = "^23.3.0"
django = "^4.2.3"
django-filter = "^23.1"
django-cors-headers = "^4.0.0"
django-split-settings = "^1.2.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
flake8 = "^6.0.0"
graphene-django = "^3.1.2"
google-auth = "^2.20.0"
google-auth-oauthlib = "^1.0.0"
google-auth-httplib2 = "^0.1.0"
google-api-python-client = "^2.90.0"
model-bakery = "^1.12.0"
python = "^3.11"
pytest = "^7.3.2"
pytest-xdist = "^3.3.1"
pytest-django = "^4.5.2"
psycopg2 = "^2.9.6"
pillow = "^9.5.0"
pyyaml = "^6.0"
python-dotenv = "^1.0.0"
boto3 = "^1.28.9"
 
[tool.poetry.extras]
argon2 = ["argon2-cffi"]

[tool.poetry.group.dev.dependencies]
colorlog = "^6.7.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.project.settings"
django_find_project = false
python_files = "test_*.py"
testpaths = ["core"]
filterwarnings = "ignore::DeprecationWarning:^(?!node\\.).*:"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
